+++ source/extensions/lightspeed.trex.app.style/lightspeed/trex/app/style/trex_style.py
@@ -419,6 +419,11 @@ current_dict.update(
         "Image::More": {"image_url": _get_icons("ellipsis"), "color": _WHITE_60},
         "Image::More:hovered": {"image_url": _get_icons("ellipsis"), "color": _WHITE_100},
         "Image::Hourglass": {"image_url": _get_icons("hourglass"), "color": _WHITE_100},
+        "Image::Particle": {"image_url": _get_icons("particle"), "color": _WHITE_60},
+        "Image::Particle:hovered": {"image_url": _get_icons("particle"), "color": _WHITE_100},
+        "Image::DeleteParticle": {"image_url": _get_icons("particle-delete"), "color": _WHITE_60},
+        "Image::DeleteParticle:hovered": {"image_url": _get_icons("particle-delete"), "color": _WHITE_100},
+        "Image::ParticleDisabled": {"image_url": _get_icons("particle"), "color": _WHITE_30},
         "Image::TreePanelLinesBackground": {
             "image_url": _get_image("45deg-256x256-1px-2px-sp-black"),
             "color": _WHITE_30,